+++ engine/src/test/java/io/camunda/zeebe/engine/state/migration/to_8_4/ColumnFamilyPrefixCorrectionMigrationTest.java
@@ -18,9 +18,10 @@ import io.camunda.zeebe.db.impl.DbString;
 import io.camunda.zeebe.engine.state.message.DbMessageState;
 import io.camunda.zeebe.engine.state.migration.MigrationTaskState;
 import io.camunda.zeebe.engine.state.migration.MigrationTaskState.State;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily48Corrector;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily49Corrector;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily50Corrector;
+import io.camunda.zeebe.engine.state.migration.to_8_2.ColumnFamilyPrefixCorrectionMigration;
+import io.camunda.zeebe.engine.state.migration.to_8_2.corrections.ColumnFamily48Corrector;
+import io.camunda.zeebe.engine.state.migration.to_8_2.corrections.ColumnFamily49Corrector;
+import io.camunda.zeebe.engine.state.migration.to_8_2.corrections.ColumnFamily50Corrector;
 import io.camunda.zeebe.engine.state.mutable.MutableProcessingState;
 import io.camunda.zeebe.engine.state.signal.SignalSubscription;
 import io.camunda.zeebe.engine.util.ProcessingStateExtension;
